windowing/gbm: move drm components into subfolder
