+++ xbmc/windowing/gbm/CMakeLists.txt
@@ -1,20 +1,14 @@
+add_subdirectory(drm)
+
 set(SOURCES OptionalsReg.cpp
             WinSystemGbm.cpp
             GBMUtils.cpp
-            DRMUtils.cpp
-            DRMLegacy.cpp
-            DRMAtomic.cpp
-            OffScreenModeSetting.cpp
             WinSystemGbmEGLContext.cpp
             GBMDPMSSupport.cpp)
 
 set(HEADERS OptionalsReg.h
             WinSystemGbm.h
             GBMUtils.h
-            DRMUtils.h
-            DRMLegacy.h
-            DRMAtomic.h
-            OffScreenModeSetting.h
             WinSystemGbmEGLContext.h
             GBMDPMSSupport.h)
 
@@ -27,4 +21,4 @@ if(OPENGLES_FOUND)
   list(APPEND HEADERS WinSystemGbmGLESContext.h)
 endif()
 
-core_add_library(windowing_Gbm)
+core_add_library(windowing_gbm)
